+++ exporters/trace/jaeger/jaeger.go
@@ -41,9 +41,9 @@ const (
 	keyEventName                     = "event"
 )
 
-// NewRawExporter returns an OTel Exporter implementation that exports the
-// collected spans to Jaeger.
-func NewRawExporter(endpointOption EndpointOption) (*Exporter, error) {
+// New returns an OTel Exporter implementation that exports the collected
+// spans to Jaeger.
+func New(endpointOption EndpointOption) (*Exporter, error) {
 	uploader, err := endpointOption.newBatchUploader()
 	if err != nil {
 		return nil, err